+++ src/waldiez/containers/nodes/model/utils/predefined.ts
@@ -78,6 +78,14 @@ const predefinedAnthropicModels = [
         value: "claude-sonnet-4-5-20250929",
     },
     {
+        label: "Claude Opus 4.5",
+        value: "claude-opus-4-5-20251101",
+    },
+    {
+        label: "Claude Haiku 4.5",
+        value: "claude-haiku-4-5-20251001",
+    },
+    {
         label: "Claude Opus 4.1",
         value: "claude-opus-4-1-20250805",
     },